fix: Prevent loss of decimals for cost inputs

+++ src/services/include/ValidationManager.h
@@ -18,7 +18,7 @@ public:
     ValidationResult<std::string> validate_venue(std::string& input);
     ValidationResult<std::string> validate_city(std::string& input);
     ValidationResult<std::string> validate_date(std::string& input);
-    ValidationResult<int32_t> validate_cost(int32_t& input);
+    ValidationResult<double> validate_cost(double& input);
     ValidationResult<int> validate_id(int& input, const std::vector<Concert>& concerts);
 
 private:
